+++ packages/wallet-sdk/src/connection/WalletLinkHTTP.ts
@@ -1,4 +1,13 @@
-class WalletLinkHTTP {
+import { ServerMessage, ServerMessageEvent } from './ServerMessage';
+
+export class WalletLinkHTTP {
+  constructor(
+    private readonly linkAPIUrl: string,
+    private readonly sessionId: string,
+    private readonly sessionKey: string,
+    private readonly handleIncomingEvent: (_: ServerMessage) => void
+  ) {}
+
   async fetchUnseenEventsAPI() {
     const credentials = `${this.sessionId}:${this.sessionKey}`;
     const auth = `Basic ${btoa(credentials)}`;
@@ -40,5 +49,22 @@ class WalletLinkHTTP {
     } else {
       throw new Error(`Check unseen events failed: ${response.status}`);
     }
+  }
+
+  // mark unseen events as seen
+  private markUnseenEventsAsSeen(events: ServerMessageEvent[]) {
+    const credentials = `${this.sessionId}:${this.sessionKey}`;
+    const auth = `Basic ${btoa(credentials)}`;
+
+    Promise.all(
+      events.map((e) =>
+        fetch(`${this.linkAPIUrl}/events/${e.eventId}/seen`, {
+          method: 'POST',
+          headers: {
+            Authorization: auth,
+          },
+        })
+      )
+    ).catch((error) => console.error('Unabled to mark event as failed:', error));
   }
 }
